src: simulator.py: added simulators
The simulators try to emulate the ingame implementation of script execution. Currently it is capable of simulating simply commands and jump instructions, being able to follow the flow of execution. It also has basic detection for crashes or other return conditions in scripts.

The user can give a hall of fame, and an offset to where in memory execution should be redirected. The code follows the execution starting at hall of fame, and simulates how often the execution flow lands in the specified memory region.

Due to ASLR which randomizes memory layout your code may sometimes crash, so success rates will likely never be 100%, even with optimized setups.

The simulator returns how many times the execution resulted in a success, and plots the amount of successes for each ASLR pattern. (there are 65)